source files and documentation for social network simulation
A simulation system to simulate the rumor spreading in the social network.  4 different roles are introduced in this system, simulating the process where rumor spreads and gets contradicted in a social network. A conversion probability computation inspired by Probabilistic Relational Classifier is used, which is one of the key parts of the system. The behaviors of entities are built according to several established data.